backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.5.2=py311haa95532_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
django=4.1=py311haa95532_0
django-cors-headers=4.2.0=pypi_0
djangorestframework=3.14.0=pypi_0
libffi=3.4.4=hd77b12b_0
numpy=1.25.2=pypi_0
openssl=3.0.10=h2bbff1b_0
pandas=2.0.3=pypi_0
pillow=10.0.0=pypi_0
pip=23.2.1=py311haa95532_0
python=3.11.4=he1021f5_0
python-dateutil=2.8.2=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3=pypi_0
setuptools=68.0.0=py311haa95532_0
six=1.16.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
sqlparse=0.4.4=py311haa95532_0
tk=8.6.12=h2bbff1b_0
tzdata=2023c=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py311haa95532_0
xz=5.4.2=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
